New predicate and function for addictive invert
