+++ Expressif.Testing/Functions/Numeric/NumericFunctionsTest.cs
@@ -121,90 +121,6 @@ namespace Expressif.Testing.Functions.Numeric
                 .Evaluate(value), Is.EqualTo(expected));
 
         [Test]
-        [TestCase(0, 1)]
-        [TestCase(1, 2)]
-        [TestCase(-1, 0)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Increment_Valid(object value, decimal? expected)
-            => Assert.That(new Increment().Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
-        [TestCase(0, -1)]
-        [TestCase(1, 0)]
-        [TestCase(-1, -2)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Decrement_Valid(object value, decimal? expected)
-            => Assert.That(new Decrement().Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
-        [TestCase(0, 3)]
-        [TestCase(1, 4)]
-        [TestCase(-1, 2)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Add_3_Valid(object value, decimal? expected)
-            => Assert.That(new Add(() => 3).Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
-        [TestCase(0, 6)]
-        [TestCase(1, 7)]
-        [TestCase(-1, 5)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Add_3Times2_Valid(object value, decimal? expected)
-            => Assert.That(new Add(() => 3, () => 2)
-                .Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
-        [TestCase(0, -3)]
-        [TestCase(1, -2)]
-        [TestCase(-1, -4)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Subtract_3_Valid(object value, decimal? expected)
-            => Assert.That(new Subtract(() => 3).Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
-        [TestCase(0, -6)]
-        [TestCase(1, -5)]
-        [TestCase(-1, -7)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Subtract_3Times2_Valid(object value, decimal? expected)
-            => Assert.That(new Subtract(() => 3, () => 2)
-                .Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
-        [TestCase(0, 0)]
-        [TestCase(1, 3)]
-        [TestCase(-1, -3)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Multiply_3_Valid(object value, decimal? expected)
-            => Assert.That(new Multiply(() => 3)
-                .Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
-        [TestCase(0, 0)]
-        [TestCase(1, 0.25)]
-        [TestCase(-1, -0.25)]
-        [TestCase("(null)", null)]
-        [TestCase("(empty)", null)]
-        [TestCase("(blank)", null)]
-        public void Divide_4_Valid(object value, decimal? expected)
-            => Assert.That(new Divide(() => 4)
-                .Evaluate(value), Is.EqualTo(expected));
-
-        [Test]
         [TestCase(0, null)]
         [TestCase(1, 1)]
         [TestCase(-1, -1)]
@@ -217,5 +133,19 @@ namespace Expressif.Testing.Functions.Numeric
         [TestCase("(blank)", null)]
         public void Invert_Valid(object value, decimal? expected)
             => Assert.That(new Invert().Evaluate(value), Is.EqualTo(expected));
+
+
+        [Test]
+        [TestCase(0, 0)]
+        [TestCase(1, -1)]
+        [TestCase(-1, 1)]
+        [TestCase(4, -4)]
+        [TestCase(-4, 4)]
+        [TestCase(null, null)]
+        [TestCase("(null)", null)]
+        [TestCase("(empty)", null)]
+        [TestCase("(blank)", null)]
+        public void Oppose_Valid(object value, decimal? expected)
+            => Assert.That(new Oppose().Evaluate(value), Is.EqualTo(expected));
     }
 }
